Add popup menu to PreviewPanel.
