+++ src/main/java/ui/view/listener/InertPopupMenuListener.java
@@ -1,0 +1,18 @@
+package ui.view.listener;
+
+import javax.swing.event.PopupMenuEvent;
+import javax.swing.event.PopupMenuListener;
+
+public class InertPopupMenuListener implements PopupMenuListener {
+    @Override
+    public void popupMenuWillBecomeVisible(PopupMenuEvent e) {
+    }
+
+    @Override
+    public void popupMenuWillBecomeInvisible(PopupMenuEvent e) {
+    }
+
+    @Override
+    public void popupMenuCanceled(PopupMenuEvent e) {
+    }
+}
